Updated project
Some behaviours have been better implemented, the dedale.jar has been
correctly added to the buildpath (verify if it works on other machines
as well! but it should)
Added a toMapRep method to Graphe but I don't know how to properly make
use of it

What's left to do is:
_ implement the call for help (SayHelloBehaviour's method actionHelp)
_ make the collectors initially explorers
_ add a list of the treasures to collectors and/or define strategies
_ implement a viable movement method for the Silo (initially an explorer
as well); maybe sthg like stay more or less in the middle of all
treasures?
_ add the moveToSilo behaviour (although if the silo doesn't move (much)
after exploration, maybe consider adding a Node siloPos to the collector
agents)